# Optimize Pascal implementations and add Delphi support