# Pascal Implementations Optimization and Delphi Support

This PR optimizes the existing Free Pascal implementations and adds Delphi versions for comparison.

## Changes

### Free Pascal (FPC)
- Fibonacci: Improved performance by ~15.6% (1.039s -> 876.7ms)
- Loops: Maintained similar performance with improved type safety
- Added project files (.lpi) with optimal compiler settings

### Added Delphi Support
- Fibonacci: 871.8ms
- Loops: 1.719s
- Included project files (.dproj) for reproducible builds

## Technical Improvements
- Changed Integer types to Cardinal for better performance
- Added compiler optimization directives
- Optimized parameter parsing using Val
- Added comprehensive documentation
- Maintained same algorithms for fair comparison

## Implementation Notes
- Delphi code requires Delphi compiler (tested with 2007)
- FPC code can be compiled on any platform
- Full compiler settings documented in project files
- Benchmark results from Windows platform